+++ gazebo/physics/bullet/BulletSliderJoint.cc
@@ -140,7 +140,7 @@ void BulletSliderJoint::SetDamping(int /*index*/, const double _damping)
 //////////////////////////////////////////////////
 void BulletSliderJoint::SetForce(int _index, double _force)
 {
-  if (_index < 0 || _index >= this->GetAngleCount())
+  if (_index < 0 || static_cast<unsigned int>(_index) >= this->GetAngleCount())
   {
     gzerr << "Calling BulletSliderJoint::SetForce with an index ["
           << _index << "] out of range\n";
