+++ gazebo/physics/ode/ODEHinge2Joint.cc
@@ -164,7 +164,7 @@ void ODEHinge2Joint::SetMaxForce(int _index, double _t)
 //////////////////////////////////////////////////
 void ODEHinge2Joint::SetForce(int _index, double _torque)
 {
-  if (_index < 0 || _index >= this->GetAngleCount())
+  if (_index < 0 || static_cast<unsigned int>(_index) >= this->GetAngleCount())
   {
     gzerr << "Calling BulletHingeJoint::SetForce with an index ["
           << _index << "] out of range\n";
