+++ gazebo/physics/ode/ODEHingeJoint.cc
@@ -142,7 +142,7 @@ double ODEHingeJoint::GetMaxForce(int /*index*/)
 //////////////////////////////////////////////////
 void ODEHingeJoint::SetForce(int _index, double _torque)
 {
-  if (_index < 0 || _index >= this->GetAngleCount())
+  if (_index < 0 || static_cast<unsigned int>(_index) >= this->GetAngleCount())
   {
     gzerr << "Calling ODEHingeJoint::SetForce with an index ["
           << _index << "] out of range\n";
